requirements.txt
anthropic>=0.40.0
python-jobspy
pandas
playwright
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.0.0
